# perf(stream): concurrently fetch row from storage and refill cache