--- a/src/common/src/config.rs
+++ b/src/common/src/config.rs
@@ -1153,8 +1153,8 @@ pub struct StreamingDeveloperConfig {
 
     /// Configure the system-wide cache row cardinality of hash join.
     /// For example, if this is set to 1000, it means we can have at most 1000 rows in cache.
-    #[serde(default = "default::developer::streaming_hash_join_cache_row_cardinality")]
-    pub hash_join_cache_row_cardinality: usize,
+    #[serde(default = "default::developer::streaming_hash_join_entry_state_max_rows")]
+    pub hash_join_entry_state_max_rows: usize,
 }
 
 /// The subsections `[batch.developer]`.
@@ -2119,7 +2119,8 @@ pub mod default {
             false
         }
 
-        pub fn streaming_hash_join_cache_row_cardinality() -> usize {
+        pub fn streaming_hash_join_entry_state_max_rows() -> usize {
+            // NOTE(kwannoel): This is just an arbitrary number.
             30000
         }
     }