--- a/src/stream/src/executor/hash_join.rs
+++ b/src/stream/src/executor/hash_join.rs
@@ -1481,53 +1481,17 @@ impl<K: HashKey, S: StateStore, const T: JoinTypePrimitive> HashJoinExecutor<K, 
                 let mut degree = 0;
                 let mut append_only_matched_row: Option<JoinRow<OwnedRow>> = None;
                 let mut matched_rows_to_clean = vec![];
-                let (
-                    order_key_indices,
-                    join_key_data_types,
-                    pk_indices,
-                    pk_serializer,
-                    state_table,
-                    mut degree_table,
-                ) = side_match.ht.get_table_mut_refs();
-                let decoded_key = key.deserialize(join_key_data_types)?;
-                let degrees = if side_match.need_degree_table {
-                    let degrees = Self::handle_fetch_degrees(
-                        key,
-                        join_key_data_types,
-                        degree_table.as_ref().unwrap(),
-                    )
+                let (matched_rows, mut degree_table) = side_match
+                    .ht
+                    .fetch_matched_rows_and_get_degree_table_ref(key)
                     .await?;
-                    Some(degrees)
-                } else {
-                    None
-                };
-                let table_iter = state_table
-                    .iter_with_prefix(&decoded_key, sub_range, PrefetchOptions::default())
-                    .await?;
-
                 #[for_await]
-                for (i, entry) in table_iter.enumerate() {
-                    let encoded_row = entry?;
-                    let encoded_pk = encoded_row
-                        .as_ref()
-                        .project(pk_indices)
-                        .memcmp_serialize(pk_serializer);
-                    let mut matched_row = JoinRow::new(
-                        encoded_row.into_owned_row(),
-                        degrees.as_ref().map_or(0, |d| d[i]),
-                    );
+                for matched_row in matched_rows {
+                    let (encoded_pk, mut matched_row) = matched_row?;
                     if entry_state_count <= entry_state_max_rows {
                         entry_state
                             .insert(encoded_pk, matched_row.encode(), None) // TODO(kwannoel): handle ineq key for asof join.
-                            .with_context(|| {
-                                let pk = row.project(pk_indices);
-                                format!(
-                                    "pk: {}, row: {}, state_table_id: {}",
-                                    pk.display(),
-                                    row.display(),
-                                    state_table.table_id()
-                                )
-                            })?;
+                            .with_context(|| format!("row: {}", row.display(),))?;
                         entry_state_count += 1;
                     }
 
@@ -1543,19 +1507,10 @@ impl<K: HashKey, S: StateStore, const T: JoinTypePrimitive> HashJoinExecutor<K, 
                     if join_condition_satisfied {
                         degree += 1;
                         if side_match.need_degree_table {
-                            // TODO: no need to `into_owned_row` here due to partial borrow.
-                            let old_degree = matched_row
-                                .to_table_rows(order_key_indices)
-                                .1
-                                .into_owned_row();
-
-                            matched_row.degree += 1;
-
-                            let new_degree = matched_row.to_table_rows(order_key_indices).1;
-                            degree_table
-                                .as_mut()
-                                .unwrap()
-                                .update(old_degree, new_degree);
+                            update_degree::<S, true>(
+                                degree_table.as_mut().unwrap(),
+                                &mut matched_row,
+                            );
                         }
                         if !forward_exactly_once(T, SIDE) {
                             if let Some(chunk) =