I hereby agree to the terms of the [RisingWave Labs, Inc. Contributor License Agreement](https://gist.github.com/TennyZhuang/f00be7f16996ea48effb049aa7be4d66#file-rw_cla).

## What's changed and what's your intention?

Currently we read all rows matching a key, encode them, and store them in an in-memory data structure. We add this to the cache. Then we iterate each row from the in-memory data structure, decode them. This means we have to incur IO latency of reading all rows first. We also have to incur the memory cost of buffering all matching records.

Instead of doing this, we can iterate the rows, concurrently refill cache for them. Then we don't have to wait for IO to finish for ALL rows. We also avoid OOM, since while refilling cache, if we notice the number of rows exceeds some threshold, we can stop refilling.

Further, if we tolerate inconsistency, we will revert to the old strategy of reading all records into cache first, so we will preserve that logic.

This is done for all join types in the hash join executor.

## Handling degree table matches and updates

Before this PR we:
1. Read from the match side degree table when doing cache refill (hold an immutable ref)
2. Update the match side degree table when doing cache refill.

However, in this PR we concurrently refill cache and handle matches. This means that 1&2 happen concurrently, which means we hold an immutable and mutable reference to the underlying match side degree table. This will be rejected by the borrow checker.

To solve this, we read from the match side degree table, keeping all the degrees in-memory. Only after that is complete, we start concurrently doing cache refill of the match side, and also updating degrees of match side. In this way the lifetimes of the mutable and immutable reference to match side degree table won't overlap.

In the future we can optimize it further, holding a separate cache for degrees, from their join state, since the degree state is much smaller.

## Nexmark Benchmark results

Build: https://buildkite.com/risingwave-test/nexmark-benchmark/builds/4994
Dashboards:

- [q3](http://metabase.risingwave-cloud.xyz/question/38984-nexmark-q3-blackhole-8c32g-1cn-affinity-avg-sink-output-rows-per-second-rows-s-history-thtb-3222?start_date=2024-10-30)
- [q5](http://metabase.risingwave-cloud.xyz/question/46071-nexmark-q5-blackhole-8c32g-1cn-affinity-avg-sink-output-rows-per-second-rows-s-history-thtb-3228?start_date=2024-11-28)
- [q7](http://metabase.risingwave-cloud.xyz/question/43834-nexmark-q7-blackhole-8c32g-1cn-affinity-avg-sink-output-rows-per-second-rows-s-history-thtb-3240?start_date=2024-11-14)
- [q8](http://metabase.risingwave-cloud.xyz/question/43905-nexmark-q8-blackhole-8c32g-1cn-affinity-avg-sink-output-rows-per-second-rows-s-history-thtb-3246?start_date=2024-11-14)
- [q20](http://metabase.risingwave-cloud.xyz/question/44231-nexmark-q20-blackhole-8c32g-1cn-affinity-avg-sink-output-rows-per-second-rows-s-history-thtb-3213?start_date=2024-11-15)
- [q101](http://metabase.risingwave-cloud.xyz/question/40466-nexmark-q101-blackhole-8c32g-1cn-affinity-avg-sink-output-rows-per-second-rows-s-history-thtb-3170?start_date=2024-10-30)
- [q102](http://metabase.risingwave-cloud.xyz/question/44256-nexmark-q102-blackhole-8c32g-1cn-affinity-avg-sink-output-rows-per-second-rows-s-history-thtb-3173?start_date=2024-11-15)
- [q105](http://metabase.risingwave-cloud.xyz/question/40602-nexmark-q105-blackhole-8c32g-1cn-affinity-avg-sink-output-rows-per-second-rows-s-history-thtb-3182?start_date=2024-10-30)

|        | unoptimized | optimized | remarks |
| - | - | - | - |
| q3 | 7.3k | 7.2k | |
| q5 | 45.7k | 222.8k | |
| q7 | 915 | 3.7k | |
| q8 | 7.7K | 6.5k | |
| q20 | 51.3K | 50.9K | |
| q101 | 250.7K | 283.3K | |
| q102 | 21.8K | 42.6K | |
| q105 | 7.34 | 4.62 | Tps is naturally low, fluctuating between 4-11 |

## Micro Benchmark results

### Cache Miss

With a 30,000 record limit for each key we measure the performance + memory consumption when a single record is matched against N records which are not in-cache. Here's the comparison (using in-memory statestore, inner join, sample size = 100):

Amp | Total Bytes | Total Bytes (opt) | Diff | Runtime (ns) | Runtime (ns) (opt) | Diff
-- | -- | -- | -- | -- | -- | --
20,000.00 | 3279096 | 3260364 | -0.57 | 17073601.28 | 16143637.7 | -5.45
40,000.00 | 6444552 | 4863221 | -24.54 | 35936541.66 | 29335407.29 | -18.37
200,000.00 | 31769736 | 4863221 | -84.69 | 191953280.4 | 100895451.7 | -47.44
400,000.00 | 63427368 | 4863221 | -92.33 | 387335817.9 | 185162442.9 | -52.20

You can observe how after 30,000, we don't do cache refill anymore, and so the memory usage becomes capped, and we can avoid OOM.

Do also note that because this is an in-memory statestore, we can see significant runtime improvements. I expect that with block cache enabled, we can also see similar improvements, although may not be as high.

### Cache Hit

We join ingest 64 records on the update (probe) side, to match all amplified records on the build side. Sample size is N, using in-memory state store. For memory utilization, we measure mem table usage, it's unavoidable because we have to use a barrier to synchronize the sequence of ingestion, so we can make the sequence of join deterministic, rather than interleaved from lhs and rhs of join. But it's fine because we don't expect better memory utilization, since everything is in cache.

Cache hit results show some regression for runtime (~5-6%):

Amp | Total Bytes | Total Bytes (opt) | Diff | Runtime (ns) | Runtime (ns) (opt) | Diff
-- | -- | -- | -- | -- | -- | --
20,000.00 | 8586702 | 8586702 | 0.00 | 225010070 | 239153682.1 | 6.29
40,000.00 | 17024222 | 17024222 | 0.00 | 457963188.8 | 482480081.3 | 5.35
200,000.00 | 84527806 | 84527806 | 0.00 | 2318848400 | 2468701041 | 6.46
400,000.00 | 168909854 | 168909854 | 0.00 | 4621595553 | 4900965761 | 6.04

However:
1. nexmark e2e performance tests show that the performance is good in most cases.
2. The microbenchmark is a highly specific case. It's an inner join with records in-cache, under high join amplification. This is not exactly the common case. Hence I think it's acceptable to further optimize later where needed.
3. The regression might be due to abstracting various functionality into standalone functions. Some of these are functions that return generators (streams), and are difficult for the compiler to inline. We can inline them by hand to improve performance. But I think we should optimize in other areas first (e.g. append_n (~50%), decode (15%)), which are major cost centers in the flamegraph, rather than optimize areas which lead to poorer readability and maintainability of the code.
4. There are also other more "obvious" optimizations, such as inlining cond expr evaluation if there's no cond expr, eliminating the `is_append_only` code path if join is non-append-only.

## Further improvements

We can also check the cache size when inserting the update records into cache. Currently we only do so for the match side, since that has N matches for a single update record. We differ this optimization to later to avoid further complicating this PR.

We may also need to call try_flush more frequently. To ensure records in join state get flushed.

We also need to remove merge chunks triggered OOM.

<!--

**Please do not leave this empty!**

Please explain **IN DETAIL** what the changes are in this PR and why they are needed:

- Summarize your change (**mandatory**)
- How does this PR work? Need a brief introduction for the changed logic (optional)
- Describe clearly one logical change and avoid lazy messages (optional)
- Describe any limitations of the current code (optional)
- Refer to a related PR or issue link (optional)

-->

## Checklist

- [ ] I have written necessary rustdoc comments
- [ ] I have added necessary unit tests and integration tests
- [ ] I have added test labels as necessary. See [details](https://github.com/risingwavelabs/risingwave/blob/main/docs/developer-guide.md#ci-labels-guide).
- [ ] I have added fuzzing tests or opened an issue to track them. (Optional, recommended for new SQL features #7934).
- [ ] My PR contains breaking changes. (If it deprecates some features, please create a tracking issue to remove them in the future).
- [ ] All checks passed in `./risedev check` (or alias, `./risedev c`)
- [ ] My PR changes performance-critical code. (Please run macro/micro-benchmarks and show the results.)
<!-- To manually trigger a benchmark, please check out [Notion](https://www.notion.so/risingwave-labs/Manually-trigger-nexmark-performance-dashboard-test-b784f1eae1cf48889b2645d020b6b7d3). -->
- [ ] My PR contains critical fixes that are necessary to be merged into the latest release. (Please check out the [details](https://github.com/risingwavelabs/risingwave/blob/main/CONTRIBUTING.md))

## Documentation

- [ ] My PR needs documentation updates. (Please use the **Release note** section below to summarize the impact on users)

## Release note

If this PR includes changes that directly affect users or other significant modifications relevant to the community, kindly draft a release note to provide a concise summary of these changes. Please prioritize highlighting the impact these changes will have on users.


<!--
Please create a release note for your changes.

Discuss technical details in the "What's changed" section, and
focus on the impact on users in the release note.

You should also mention the environment or conditions where the impact may occur.
-->

</details>