@fuyufjh 2024-12-04
The motivaion is similiar to #10979 / #12615, but this is much better. Love it!

IIRC, Nexmark has some queries with large join state. May use it as benchmark.