@kwannoel 2024-12-09
Added benchmark: https://github.com/risingwavelabs/risingwave/pull/19712. Continuing investigation.