@kwannoel 2024-12-11
We don't have any issues supporting this for `INNER JOIN`, however for joins which require a degree table, more consideration is needed. Carving out this optimization just for `INNER JOIN` is possible, but it will add a new code branch and more complexity.

---

Currently we:
1. Read all the degrees and matched rows into memory.
2. Iterate through them, updating the degree table for the build side.

So the upfront IO cost is: `max(read_degree_latency, read_matched_rows_latency)`. Typically `read_matched_rows_latency` should be higher, since `degree` will have less data, only key + degree.

Then the total latency for cache missed will be `max(read_degree_latency, read_matched_rows_latency) + cpu latency of processing matched rows`.

After this PR:
We concurrently read **matched rows** and handle them. However, for **degrees** we can't do so because
we need to iterate over the degrees, and update them at the same time. This means holding an immutable and mutable reference to `StateTable` at the same time.

Currently I workaround this by reading all the degrees into memory first. The size of the degrees shouldn't be too bad, since we just need to store the degrees.
This means the upfront IO cost will be `read_degree_latency`. Then we will need to pay IO costs of `read_matched_rows_latency` in an amortized way, since we concurrently stream and handle each matched row.

Then the total latency for cache missed will be `read_degree_latency + cpu AND io latency of processing matched rows`.
So it *might* take longer to do cache refill.

However, given that cache miss is expected to seldom occur, I think this is a reasonable trade-off to make.

Will benchmark this approach against a nexmark query with `LEFT OUTER JOIN`.

Another approach is to concurrently do point get, and write to the degree state table, since the point get reference to the degree table is short-lived. However, point get does not have prefetch interface yet.

---

The second issue is tolerating inconsistency. Previously we buffered all matched rows into memory. If there's a mismatch in the number of rows in the degree table rows and the matched table rows, we will compare their pk.

I think it's possible to support this with the point get approach. Alternatively, if tolerate inconsistency is set to true, we can always follow the old approach of refilling cache first.