@kwannoel 2024-12-27
Benchmarking https://buildkite.com/risingwave-test/nexmark-benchmark/builds/4974