@st1page 2025-01-06
> We concurrently read matched rows and handle them. However, for degrees we can't do so because
we need to iterate over the degrees, and update them at the same time. This means holding an immutable and mutable reference to StateTable at the same time.
> Currently I workaround this by reading all the degrees into memory first. The size of the degrees shouldn't be too bad, since we just need to store the degrees.

I haven't read the code closely enough. But there are still some contention issues here after just reading all the degree into memory? For a chunk of input, if there are multiple rows corresponding to the same join key, they should access the join key's degree in order. 🤔