@kwannoel 2025-01-07
> > We concurrently read matched rows and handle them. However, for degrees we can't do so because
> > we need to iterate over the degrees, and update them at the same time. This means holding an immutable and mutable reference to StateTable at the same time.
> > Currently I workaround this by reading all the degrees into memory first. The size of the degrees shouldn't be too bad, since we just need to store the degrees.
> 
> ~ I haven't read the code closely enough. But there are still some contention issues here after just reading all the degree into memory? For a chunk of input, if there are multiple rows corresponding to the same join key, they should access the join key's degree in order. 🤔 ~
> 
> Nevermind, I think it wrong(

Good question, I also thought about it before. Let me explain here, in case others have the same question. We process the matches for each record in an update chunk one by one. For each record, we will fetch the matched rows and their degrees from cache or object store. Even if multiple records in an update chunk have the same join key, each match for these records will fetch consistent state. Degrees of the cached join rows are also updated accordingly, to make sure they are consistent after matching.